Consolidate coercion rules
